Creacion de parte de comentarios y usuarios, repositorio, controladores, servicios

+++ src/main/java/com/saborurbano/restaurante/model/Comentario.java
@@ -16,7 +16,7 @@ public class Comentario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idComentario;
+    private int idComentario;
 
     @Column(columnDefinition = "TEXT")
     private String textoComentario;
